dvcdbg

Module scanner

Functions

scan_i2c

Scan the I2C bus for connected devices (addresses 0x03 to 0x77).

scan_i2c_with_ctrl

Scan the I2C bus for devices by sending specified control bytes.

scan_init_sequence

Scan the I2C bus by testing each command in an initialization sequence.