dvcdbg::scanner

Function scan_i2c

pub fn scan_i2c<I2C, L>(i2c: &mut I2C, logger: &mut L)
where
    I2C: I2c,
    L: Logger,

Scan the I2C bus for connected devices (addresses 0x03 to 0x77).

Arguments

i2c - Mutable reference to the I2C interface implementing embedded_hal::i2c::I2c.

logger - Mutable reference to a logger implementing the Logger trait.

This function attempts to write zero bytes to each possible device address on the I2C bus, logging addresses that respond successfully.